# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: 通学AIラジオ API
  version: "0.1.0"
  description: |
    フロントエンド（React + Vite）とバックエンド（Go）を並行開発するための取り決め。
    **実装より先にこのファイルを更新して合意すること。** 変更したら両チームに共有する。

    - データモデルは [06-storage.md](../backend/docs/pipeline/06-storage.md) の PostgreSQL スキーマに対応する
    - 認証は行わない（ログイン機能はスコープ外。`userId` の発行方法は下記「userId の扱い」参照）。
      ローカル完結・オフライン前提の構成であり、本番公開は想定していないための割り切り

    ## CORS
    フロントエンド（`http://localhost:3000` 想定）とバックエンド（`http://localhost:8080` 想定）は
    オリジンが異なるため、バックエンドは以下の CORS レスポンスヘッダを返すこと。
    実装時に抜けやすく、抜けると疎通確認の最初の一歩で詰まるので明記しておく。

    ```
    Access-Control-Allow-Origin: http://localhost:3000
    Access-Control-Allow-Methods: GET, POST, PUT, OPTIONS
    Access-Control-Allow-Headers: Content-Type
    ```

    ## userId の扱い
    ログイン機能はスコープ外のため、次の方式で暫定的にユーザーを識別する。

    1. フロントエンドは初回起動時に `POST /api/users` を呼び、サーバーが発行した `userId`（UUID）を受け取る
    2. `userId` をブラウザの `localStorage` に保存し、以降の全リクエストで使い回す
    3. サーバーはリクエストされた `userId` が `users` テーブルに存在するかを検証し、
       存在しなければ `404`（`USER_NOT_FOUND`）を返す

    クライアントが自前で UUID を生成しない理由：`programs.user_id` が外部キー制約付きのため、
    存在しないIDでの書き込みを防ぐにはサーバー発行が確実。

    ## 表示名（ニックネーム）の扱い
    UIモック（`/mock`）ではプロフィール画面にユーザー名を表示するが、ログイン機能がスコープ外である以上
    サーバー側でアカウントに紐づく表示名を管理する意味は薄い。表示名はバックエンドAPIの対象外とし、
    フロントエンドの `localStorage` にのみ保存するクライアントローカルな値として扱う
    （どのAPIレスポンスにも `name` 相当のフィールドは含まれない）。
servers:
  - url: http://localhost:8080
    description: ローカル開発環境（VITE_API_BASE_URL、詳細は frontend/CLAUDE.md#環境変数）
...
paths:
...
  /api/users/{userId}/programs/latest:
    get:
      operationId: getLatestProgram
      tags: [programs]
      summary: 最新の番組を取得する（ホーム）
      description: |
        その日（最新）の番組を取得する（[ホーム](../frontend/docs/features/home.md)）。
        バッチ処理（`POST /api/batch/run`）が生成した、その `userId` にとって
        最も新しい番組を返す。
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "404":
          description: "userId が存在しない、または番組がまだ1件も生成されていない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                userNotFound:
                  summary: userId が存在しない
                  value:
                    error:
                      code: USER_NOT_FOUND
                      message: "指定されたユーザーが見つかりません"
                programNotFound:
                  summary: 番組が未生成（ホーム画面は「準備中」表示の分岐に使う）
                  value:
                    error:
                      code: PROGRAM_NOT_FOUND
                      message: "指定されたユーザーの番組がまだ生成されていません"
  /api/users/{userId}/programs:
    get:
      operationId: listPrograms
      tags: [programs]
      summary: 過去の番組一覧を取得する（ホームの履歴）
      description: |
        ホーム画面の「履歴」セクション用。作成日時の降順で返す。個別記事の全文検索や
        ビッグデータ的な分析は行わない（[Requirements.md](../Requirements.md)のスコープ外参照）、
        あくまで直近の番組メタデータの一覧に限定した軽量なエンドポイント。

        各要素の `title` は、その番組内で最も重要度が高いチャプター（`position: 0`）の
        `title` をサーバー側でそのまま採用したもの（新たに番組全体の要約タイトルを
        生成するわけではない）。
      parameters:
        - $ref: "#/components/parameters/UserId"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
          description: 取得件数の上限
      responses:
        "200":
          description: "取得成功（1件も無ければ `items` は空配列）"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramHistoryResponse"
        "404":
          description: "`userId` が存在しない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: USER_NOT_FOUND
                  message: "指定されたユーザーが見つかりません"
...
  /api/programs/{programId}:
    get:
      operationId: getProgram
      tags: [programs]
      summary: 指定した番組の詳細を取得する（プレイヤー）
      description: |
        指定した番組の詳細を取得する（[プレイヤー](../frontend/docs/features/player.md)）。
        レスポンス形式は `GET /api/users/{userId}/programs/latest` と同じ
        （`userId` に紐付かない直接アクセス用。ホームからプレイヤーへ遷移する際に番組IDで参照する）。
      parameters:
        - $ref: "#/components/parameters/ProgramId"
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "404":
          description: 番組が存在しない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: PROGRAM_NOT_FOUND
                  message: "指定された番組が見つかりません"
...
components:
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProgramId:
      name: programId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Program:
      type: object
      required: [id, title, createdAt, greetingText, changeCount, totalDurationSec, chapters]
      properties:
        id:
          type: string
          format: uuid
          example: 8f14e45f-ceea-4c9e-8b52-1c4d5d5f6a12
        title:
          type: string
          description: |
            番組の代表タイトル。最も重要度が高いチャプター（`position: 0`）の`title`を
            サーバー側でそのまま流用したもの（番組全体を要約する専用のタイトルを
            別途生成するわけではない）。ホームの履歴一覧・プレイヤーの見出しに使う
          example: "AI開発の新モデルが発表、性能が前世代比2倍に"
        createdAt:
          type: string
          format: date-time
          example: "2026-08-29T06:00:00+09:00"
        greetingText:
          type: string
          description: |
            番組冒頭の挨拶文（台本の書き出しと同じ文面）。ホーム画面の
            「ゲストさん、おはようございます。」に続けて表示する本文パート。
            ユーザーの表示名はクライアントローカルの値のため、このテキストには含まない
            （本ファイル冒頭「表示名（ニックネーム）の扱い」参照）
          example: "今日はAI関連で3件、動きがあります。"
        changeCount:
          type: integer
          description: |
            前日の番組からの差分件数（新規・更新されたトピック数）。`chapters.length`とは
            別概念。`demo/generate`以外（キャッシュ経由の生成）では
            [03-selection.md](../backend/docs/generation/03-selection.md)の選択ロジックが、
            `demo/generate`（ライブfetch経路）では重要度判定ステップ
            （[03-score.md](../backend/docs/pipeline/03-score.md)）が算出する。
            ホーム画面の「今日はn件、動きがあります」の`n`に対応
          example: 3
        totalDurationSec:
          type: integer
          description: 番組合計の再生時間（秒）
          example: 420
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/Chapter"
    Chapter:
      type: object
      required:
        - id
        - position
        - title
        - sourceUrl
        - sourceName
        - script
        - audioUrl
        - durationSec
        - lineStartOffsetsSec
      properties:
        id:
          type: string
          format: uuid
          example: c56a4180-65aa-42ec-a945-5fd21dec0538
        position:
          type: integer
          description: 番組内の並び順
          example: 0
        title:
          type: string
          example: "〇〇について"
        sourceUrl:
          type: string
          format: uri
          example: https://example.com/article
        sourceName:
          type: string
          example: Example News
        script:
          type: string
          description: |
            このチャプターの読み上げ台本全文（[04-script.md](../backend/docs/pipeline/04-script.md)で
            生成されたもの、`chapters.script`カラムに対応）。改行区切りの複数行からなる。
            プレイヤー画面でSpotifyの歌詞表示のように現在の読み上げ箇所をハイライトする用途に使う
            （各行の実際の発話開始位置は`lineStartOffsetsSec`を参照）。
          example: "まず注目は新しい言語モデルの発表です。\n前世代と比べて推論速度が2倍になったとのことです。"
        audioUrl:
          type: string
          example: /api/audio/8f14e45f-ceea-4c9e-8b52-1c4d5d5f6a12/c56a4180-65aa-42ec-a945-5fd21dec0538
        durationSec:
          type: integer
          description: |
            チャプター単体の再生時間（秒）。プレイヤーのシークバー初期表示に使う
            （`<audio>`のメタデータ読み込み完了を待たずに表示できるようにするため）
          example: 95
        lineStartOffsetsSec:
          type: array
          items:
            type: number
          description: |
            `script`を改行分割した各行（空行を除去する前のインデックス）が実際に発話開始する、
            チャプター先頭からの経過秒数（秒、小数）。TTS合成時
            （[05-tts.md](../backend/docs/pipeline/05-tts.md)）に話者ごとの読み上げ速度・行間ポーズを
            含めて実測した値であり、文字数比率による推定ではない。本フィールド追加前に生成された
            番組では空配列になり、クライアントはその場合に文字数比按分へフォールバックする。
          example: [0, 3.2, 7.8]
...
    ProgramSummary:
      type: object
      required: [id, title, createdAt, totalDurationSec]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: "円安基調が継続、輸出企業への影響は"
        createdAt:
          type: string
          format: date-time
        totalDurationSec:
          type: integer
          example: 660
    ProgramHistoryResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProgramSummary"
        total:
          type: integer
          description: 条件に合致する全件数（`limit`で切られる前の総数）
          example: 12
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: "`SCREAMING_SNAKE_CASE`の固定文字列"
              enum:
                - USER_NOT_FOUND
                - INVALID_TAGS
                - INVALID_SETTINGS
                - PROGRAM_NOT_FOUND
                - AUDIO_NOT_FOUND
                - ALREADY_GENERATING
                - RESET_LIMIT_EXCEEDED
                - ARTICLE_CACHE_EMPTY
                - NO_UNSEEN_ARTICLES
                - UPSTREAM_LLM_FAILED
                - UPSTREAM_TTS_FAILED
                - INTERNAL_ERROR
            message:
              type: string
...